requirements.txt
Babel==2.12.1
blinker==1.6.2
click==8.1.3
defusedxml==0.7.1
Flask==2.3.2
flask-babel==3.1.0
Flask-Login==0.6.2
Flask-OpenID==1.3.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.3.3
WTForms==3.0.1
